requirements.txt
Flask==3.1.2
itsdangerous==2.2.0
matplotlib==3.10.5
numpy==2.2.6
pandas==2.3.2
plotly==6.3.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
Werkzeug==3.1.3
openpyxl==3.1.2
gunicorn==21.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0 